+++ hoge.php
@@ -1,8 +1,0 @@
-<?
-
-define('DS', '/');
-
-include('Cake/bootstrap.php');
-
-
-?>