+++ app/HISTORY.md
@@ -10,7 +10,9 @@ Running Kingdom
 	* ユーザDBテーブルのスキーマ設計
 * cakephp用のoauthライブラリ http://code.42dh.com/oauth/
 * https://runkeeper.com/apps/authorize?client_id=fa85c607244c491f825f66e8dcf704ef&response_type=code&redirect_uri=http://vps.arifumi.net
-* 結局、ほとんどライブラリのお世話になることなく、tokenを取得できた。でもこれって、ちゃんとOAuth実装できてんのかな。
+	* 結局、ほとんどライブラリのお世話になることなく、tokenを取得できた。でもこれって、ちゃんとOAuth実装できてんのかな。
+* code: 24ae2d4bfdcc44ff90aaf66157ad56ab
+* token: d3819b92bb874523b9c0e18ba9351c51
 
 2013.5.11 at library
 --------------------